net/ionic: split datapath into simple and scatter-gather versions
The code is very similar, but the simple case can skip a few branches
in the hot path.

SG is enabled on the Rx side with offload DEV_RX_OFFLOAD_SCATTER.
SG is enabled on the Tx side with offload DEV_TX_OFFLOAD_MULTI_SEGS.

SG is automatically enabled on the Rx side if the provided mbufs are
too small to hold the maximum possible frame.

To enable SG in testpmd, add these args:
  --rx-offloads=0x2000 --tx-offloads=0x8000
